Move InsertAnswerTables to its own file
